+++ src/helpers/TaxonomyHelper.ts
@@ -1,5 +1,11 @@
 import { getTaxonomyField } from './getTaxonomyField';
-import { EXTENSION_NAME, SETTING_TAXONOMY_CUSTOM } from '../constants';
+import {
+  EXTENSION_NAME,
+  LocalStore,
+  SETTING_TAXONOMY_CATEGORIES,
+  SETTING_TAXONOMY_CUSTOM,
+  SETTING_TAXONOMY_TAGS
+} from '../constants';
 import { CustomTaxonomy, TaxonomyType, ContentType as IContentType } from '../models';
 import { FilesHelper } from './FilesHelper';
 import { ProgressLocation, window } from 'vscode';
@@ -11,8 +17,101 @@ import { Notifications } from './Notifications';
 import { ArticleHelper } from './ArticleHelper';
 import { ContentType } from './ContentType';
 import { readFileAsync, writeFileAsync } from '../utils';
+import { Config, JsonDB } from 'node-json-db';
+import { Folders } from '../commands';
+import { join } from 'path';
+import { SettingsListener as PanelSettingsListener } from '../listeners/panel';
+import { SettingsListener as DashboardSettingsListener } from '../listeners/dashboard';
 
 export class TaxonomyHelper {
+  private static db: JsonDB;
+
+  /**
+   * Get all the taxonomy values
+   */
+  public static async getAll() {
+    if (!this.db) {
+      const db = this.initDb();
+      if (!db) {
+        return;
+      }
+
+      this.db = db;
+    }
+
+    const taxonomyData = {
+      tags: (await TaxonomyHelper.get(TaxonomyType.Tag)) || [],
+      categories: (await TaxonomyHelper.get(TaxonomyType.Category)) || [],
+      customTaxonomy: Settings.get<CustomTaxonomy[]>(SETTING_TAXONOMY_CUSTOM) || []
+    };
+
+    return taxonomyData;
+  }
+
+  /**
+   * Get the taxonomy settings
+   *
+   * @param type
+   * @param options
+   */
+  public static async get(type: TaxonomyType): Promise<string[] | undefined> {
+    if (!this.db) {
+      const db = this.initDb();
+      if (!db) {
+        return;
+      }
+
+      this.db = db;
+    }
+
+    const tagType = TaxonomyHelper.getTaxonomyDbPath(type);
+
+    let taxonomy: string[] = [];
+    if (await this.db.exists(tagType)) {
+      taxonomy = await this.db.getObject<string[]>(tagType);
+    }
+    return taxonomy;
+  }
+
+  /**
+   * Update the taxonomy settings
+   *
+   * @param type
+   * @param options
+   */
+  public static async update(type: TaxonomyType, options: string[]) {
+    if (!this.db) {
+      const db = this.initDb();
+      if (!db) {
+        return;
+      }
+
+      this.db = db;
+    }
+
+    const tagType = TaxonomyHelper.getTaxonomyDbPath(type);
+
+    options = [...new Set(options)];
+    options = options.sort().filter((o) => !!o);
+
+    await this.db.push(tagType, options, true);
+
+    // Trigger the update of the taxonomy
+    PanelSettingsListener.getSettings();
+    DashboardSettingsListener.getSettings(true);
+  }
+
+  /**
+   * Get the Taxonomy path of the db entry
+   * @param type
+   * @returns
+   */
+  public static getTaxonomyDbPath(type: TaxonomyType) {
+    let tagType = type === TaxonomyType.Tag ? SETTING_TAXONOMY_TAGS : SETTING_TAXONOMY_CATEGORIES;
+    tagType = tagType.replace('.', '/');
+    return `/${tagType}`;
+  }
+
   /**
    * Rename an taxonomy value
    * @param data
@@ -56,7 +155,7 @@ export class TaxonomyHelper {
 
     let options = [];
     if (taxonomyType === TaxonomyType.Tag || taxonomyType === TaxonomyType.Category) {
-      options = Settings.getTaxonomy(taxonomyType);
+      options = (await TaxonomyHelper.get(taxonomyType)) || [];
     } else {
       options = Settings.getCustomTaxonomy(taxonomyType);
     }
@@ -114,7 +213,7 @@ export class TaxonomyHelper {
     const { type } = data;
 
     const taxonomyType = this.getTypeFromString(type);
-    const options = this.getTaxonomyOptions(taxonomyType);
+    const options = await this.getTaxonomyOptions(taxonomyType);
 
     const newOption = await window.showInputBox({
       title: `Create a new ${type} value`,
@@ -367,6 +466,28 @@ export class TaxonomyHelper {
   }
 
   /**
+   * Initialize the database
+   * @returns
+   */
+  private static initDb() {
+    const wsFolder = Folders.getWorkspaceFolder();
+    if (!wsFolder) {
+      return;
+    }
+
+    const dbFolder = join(
+      parseWinPath(wsFolder?.fsPath || ''),
+      LocalStore.rootFolder,
+      LocalStore.databaseFolder
+    );
+    const dbPath = join(dbFolder, LocalStore.taxonomyDatabaseFile);
+
+    const db = new JsonDB(new Config(dbPath, true, false, '/'));
+
+    return db;
+  }
+
+  /**
    * Retrieve the fields for the taxonomy field
    * @returns
    */
@@ -413,7 +534,7 @@ export class TaxonomyHelper {
     newValue?: string
   ) {
     // Update the settings
-    let options = this.getTaxonomyOptions(taxonomyType);
+    let options = await this.getTaxonomyOptions(taxonomyType);
 
     const idx = options.findIndex((o) => o === oldValue);
     if (newValue) {
@@ -429,7 +550,7 @@ export class TaxonomyHelper {
     }
 
     if (taxonomyType === TaxonomyType.Tag || taxonomyType === TaxonomyType.Category) {
-      await Settings.updateTaxonomy(taxonomyType, options);
+      TaxonomyHelper.update(taxonomyType, options);
     } else {
       await Settings.updateCustomTaxonomyOptions(taxonomyType, options);
     }
@@ -440,11 +561,11 @@ export class TaxonomyHelper {
    * @param taxonomyType
    * @returns
    */
-  private static getTaxonomyOptions(taxonomyType: TaxonomyType | string) {
+  private static async getTaxonomyOptions(taxonomyType: TaxonomyType | string) {
     let options = [];
 
     if (taxonomyType === TaxonomyType.Tag || taxonomyType === TaxonomyType.Category) {
-      options = Settings.getTaxonomy(taxonomyType);
+      options = (await TaxonomyHelper.get(taxonomyType)) || [];
     } else {
       options = Settings.getCustomTaxonomy(taxonomyType);
     }